Refactoring & Improvement

+++ test/lib/test_common.bash
@@ -2,8 +2,6 @@
 
 source lib/common.bash
 
-is_executed && return
-
 TEST_DIR="${REPO_DIR}/test"
 
 _archive_repo() {
@@ -21,6 +19,7 @@ _build_image() {
     info "Building docker image ${image_name}"
     docker buildx build --pull --load \
         --build-arg bootstrap="${bootstrap}" \
+        --build-arg dry_run="${dry_run}" \
         -t "${image_name}" \
         -f "${dockerfile}" \
         .
@@ -40,6 +39,7 @@ _run_container() {
         --env ARCHIVE_URL="file:///archive.tar.gz" \
         --env NO_PROTO_CHECK=true \
         --env NO_GIT_CLONE=true \
+        --env GROUP_PREFIX="$(show_current_group)" \
         --name "${container_name}" \
         "${image_name}" \
         sleep infinity
@@ -59,11 +59,27 @@ _exec_in_container() {
     info "Executing \`$*\` in docker container ${container_name}"
     docker exec \
         --env GITHUB_ACTIONS \
+        --env GROUP_PREFIX="$(show_current_group)" \
+        "${container_name}" \
+        "$@"
+}
+
+_exec_in_container_root() {
+    local -r container_name="${1}"
+    shift
+
+    info "Executing \`$*\` in docker container ${container_name}"
+    docker exec \
+        --env GITHUB_ACTIONS \
+        --env GROUP_PREFIX="$(show_current_group)" \
+        --user root:root \
         "${container_name}" \
         "$@"
 }
 
 test_bootstrap() {
+    local -r bootstrap=true
+    local -r dry_run=false
     local -r test_os_name="${1}"
     shift
 
@@ -85,7 +101,7 @@ test_bootstrap() {
 
     group_start "Setup container"
     {
-        _build_image "${image_name}" "${dockerfile}" true
+        _build_image "${image_name}" "${dockerfile}" "${bootstrap}" "${dry_run}"
         _run_container "${image_name}" "${container_name}" "${archive_path}"
     }
     group_end
@@ -94,6 +110,7 @@ test_bootstrap() {
     {
         # Run the bootstrap script
         _exec_in_container "${container_name}" bash -c "cat /bootstrap.bash | bash -s -- $*"
+        reset_last_group # The bootstrap script may output group_end
     }
     group_end
 
@@ -101,16 +118,21 @@ test_bootstrap() {
     {
         _stop_container "${container_name}"
     }
+    group_end
 
     group_start "Cleanup working directory"
     {
         rm -rf "${work_dir}"
     }
+    group_end
 
     info "Test for ${test_os_name} succeeded"
 }
 
-test_install() {
+_test_install_common() {
+    local -r bootstrap=false
+    local -r dry_run="${1}"
+    shift
     local -r test_os_name="${1}"
     shift
 
@@ -132,44 +154,62 @@ test_install() {
 
     group_start "Setup container"
     {
-        _build_image "${image_name}" "${dockerfile}"
+        _build_image "${image_name}" "${dockerfile}" "${bootstrap}" "${dry_run}"
         _run_container "${image_name}" "${container_name}" "${archive_path}"
     }
     group_end
 
     group_start "Extract archive"
     {
-        _exec_in_container "${container_name}" sudo mkdir -p /dotfiles
+        _exec_in_container_root "${container_name}" mkdir -p /dotfiles
         local uid gid
         uid="$(_exec_in_container "${container_name}" id -u)"
         gid="$(_exec_in_container "${container_name}" id -g)"
-        _exec_in_container "${container_name}" sudo chown -R "${uid}:${gid}" /dotfiles
+        _exec_in_container_root "${container_name}" chown -R "${uid}:${gid}" /dotfiles
         _exec_in_container "${container_name}" tar -xzf /archive.tar.gz -C /
     }
     group_end
 
     group_start "Run the install script"
     {
-        # Run the bootstrap script
+        # Run the install script
         _exec_in_container "${container_name}" bash -c "/dotfiles/install $*"
+        reset_last_group # The install script may output group_end
     }
+    group_end
 
-    # Run the test script again to make sure it's idempotent
-    group_start "Run the install script again"
-    {
-        # Run the bootstrap script
-        _exec_in_container "${container_name}" bash -c "/dotfiles/install $*"
-    }
+    if [[ "${dry_run}" = false ]]; then
+        # Run the test script again to make sure it's idempotent
+        group_start "Run the install script again"
+        {
+            # Run the bootstrap script
+            _exec_in_container "${container_name}" bash -c "/dotfiles/install $*"
+            reset_last_group # The install script may output group_end
+        }
+        group_end
+    fi
 
     group_start "Cleanup container"
     {
         _stop_container "${container_name}"
     }
+    group_end
 
     group_start "Cleanup working directory"
     {
         rm -rf "${work_dir}"
     }
+    group_end
 
     info "Test for ${test_os_name} succeeded"
+}
+
+test_install_normal() {
+    local dry_run=false
+    _test_install_common "${dry_run}" "$@"
+}
+
+test_install_dry_run() {
+    local dry_run=true
+    _test_install_common "${dry_run}" "$@" --dry-run
 }
